Introduce Admin::Tables namespace
