--- app/components/admin/table_component.rb
@@ -1,4 +1,0 @@
-class Admin::TableComponent < ApplicationComponent
-  renders_many :headers, Admin::TableHeaderComponent
-  renders_many :rows, Admin::TableRowComponent
-end
