--- app/components/admin/table_row_component.rb
@@ -1,3 +1,0 @@
-class Admin::TableRowComponent < ApplicationComponent
-  renders_many :cells, Admin::TableCellComponent
-end
