+++ app/components/admin/tables/row_component.html.erb
