+++ app/components/admin/tables/table_component.html.erb
