+++ app/components/admin/tables/table_component.rb
@@ -1,0 +1,8 @@
+module Admin
+  module Tables
+    class TableComponent < ApplicationComponent
+      renders_many :headers, HeaderComponent
+      renders_many :rows, RowComponent
+    end
+  end
+end
