+++ app/views/admin/transactions/index.html.erb
@@ -2,7 +2,7 @@
   <%= render Admin::HeadingComponent.new(t(".title")) %>
 
   <%= render Admin::ContentComponent.new do %>
-    <%= render Admin::TableComponent.new do |table| %>
+    <%= render Admin::Tables::TableComponent.new do |table| %>
       <% table.with_header "Date" %>
       <% table.with_header "Type" %>
       <% table.with_header "Description" %>
